+binaryToCharacter()
added function that reads int array representing binary value and returing its associated character. also advanced embedSecretMessage function to read the LSB we add to the byte and storing it in an int array, after 8LSBs we have 1 byte of binary data that represents the character we just embedded. Passing this array to the new function binaryToCharacter returns as the character we just embedded. we printing that character