decompressing
Added -d functionality. It can now extract the message that is compressed in the image done by same programs -c command. It reads all the bytes in the image and scans over all the 8bytes LSBs to form 1 byte of message character data. Storing it in an array were the secret message is formed until character '~'. ohyeah also advanced the readMessage() to add the '~' character at the end so it gets embedded at the end of the secretMessage in the image.